requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
tidb-vector>=0.0.9
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Embedding models (all optional - install only what you need)
openai>=1.12.0  # Optional: for OpenAI embeddings
sentence-transformers>=2.2.0  # Optional: for HuggingFace embeddings
transformers>=4.30.0  # Optional: for HuggingFace embeddings
torch>=2.0.0  # Optional: for HuggingFace embeddings
ollama>=0.1.0  # Optional: for locally hosted Ollama models (Qwen, Llama, etc.)

# Evaluation and utilities
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
